chore(origin-ui-core): move selected cert in bundleItemEdit
